+++ SystemControl/startmodule_patch.c
@@ -164,7 +164,7 @@ static int _sceKernelStartModule(int modid, SceSize argsize, void *argp, int *mo
 				ret = get_game_tag(path, &tag);
 
 				if (ret == 0 && tag != (u32)-1) {
-					printk("%s: tag 0x%08X\n", __func__, tag);
+					printk("%s: tag 0x%08X\n", __func__, (uint)tag);
 					load_opnssmp(path, tag);
 				} else {
 					printk("%s: get_game_tag -> 0x%08X\n", __func__, ret);
